+++ src/Monq.Models.Abstractions/DateRangePostViewModel.cs
@@ -12,13 +12,13 @@ namespace Monq.Models.Abstractions
         /// </summary>
         [Required(ErrorMessage = "Не указана начальная точка диапазона дат.")]
         [Range(default, 2147483647, ErrorMessage = "Недопустимое значение начальной точки диапазона дат в формате Unixtimestamp.")]
-        public int Start { get; set; }
+        public long Start { get; set; }
 
         /// <summary>
         /// Конечная точка диапазона.
         /// </summary>
         [Required(ErrorMessage = "Не указана конечная точка диапазона дат.")]
         [Range(default, 2147483647, ErrorMessage = "Недопустимое значение конечной точки диапазона дат в формате Unixtimestamp.")]
-        public int End { get; set; }
+        public long End { get; set; }
     }
 }
